Add test for packet sequence
